requirements.txt
imbalanced_learn==0.11.0
matplotlib==3.7.2
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
scikit_learn==1.3.0
seaborn==0.12.2
streamlit==1.25.0
